+++ www/src/utils/node/docs.js
@@ -35,9 +35,6 @@ exports.createPages = async ({ graphql, actions }) => {
   const contributorPageTemplate = path.resolve(
     `src/templates/template-contributor-page.js`
   )
-  const localPackageTemplate = path.resolve(
-    `src/templates/template-docs-local-packages.js`
-  )
 
   const { data, errors } = await graphql(`
     query {
@@ -53,7 +50,6 @@ exports.createPages = async ({ graphql, actions }) => {
           fields {
             slug
             locale
-            package
             released
           }
           frontmatter {
@@ -193,15 +189,6 @@ exports.createPages = async ({ graphql, actions }) => {
             ...prevAndNext,
           },
         })
-      } else if (node.fields.package) {
-        // Local package template
-        createPage({
-          path: `${node.fields.slug}`,
-          component: slash(localPackageTemplate),
-          context: {
-            slug: node.fields.slug,
-          },
-        })
       } else {
         // Docs template
         createPage({
@@ -280,20 +267,6 @@ exports.onCreateNode = ({ node, actions, getNode }) => {
       }
     }
 
-    // Add slugs for package READMEs.
-    if (
-      fileNode.sourceInstanceName === `packages` &&
-      parsedFilePath.name === `README`
-    ) {
-      locale = "en"
-      slug = `/packages/${parsedFilePath.dir}/`
-      createNodeField({
-        node,
-        name: `title`,
-        value: parsedFilePath.dir,
-      })
-      createNodeField({ node, name: `package`, value: true })
-    }
     if (slug) {
       createNodeField({ node, name: `anchor`, value: slugToAnchor(slug) })
       createNodeField({ node, name: `slug`, value: slug })