+++ packages/gatsby-theme-blog/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.21](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog@1.5.20...gatsby-theme-blog@1.5.21) (2020-05-08)
+
+**Note:** Version bump only for package gatsby-theme-blog
+
+## [1.5.20](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog@1.5.19...gatsby-theme-blog@1.5.20) (2020-05-08)
+
+**Note:** Version bump only for package gatsby-theme-blog
+
+## [1.5.19](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog@1.5.18...gatsby-theme-blog@1.5.19) (2020-05-07)
+
+**Note:** Version bump only for package gatsby-theme-blog
+
+## [1.5.18](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog@1.5.17...gatsby-theme-blog@1.5.18) (2020-05-07)
+
+**Note:** Version bump only for package gatsby-theme-blog
+
+## [1.5.17](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog@1.5.16...gatsby-theme-blog@1.5.17) (2020-05-07)
+
+**Note:** Version bump only for package gatsby-theme-blog
+
 ## [1.5.16](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog@1.5.15...gatsby-theme-blog@1.5.16) (2020-05-06)
 
 **Note:** Version bump only for package gatsby-theme-blog